+++ bin/aws-must.rb
@@ -53,12 +53,14 @@ class App < Thor
 
   # ------------------------------------------------------------------
   add_shared_option :template_gem, :aliases => "-g", :type => :string,
-  :desc => "Mustache template Gem name (optionally with version constraint e.g. 'aws-must-templates,~>0.0.1')"
+  :desc => "Mustache template Gem name (optionally with version constraint e.g. 'aws-must-templates,~>0.0.1') (DEPRECATED)"
 
   add_shared_option :template_path, :aliases => "-t", :type =>
-  :string, :default => "mustache", 
-  :desc => "Directory holding mustache templates"
+  :string, :desc => "Directory holding mustache templates (DEPRECATED)"
 
+  add_shared_option :mustaches, :aliases => "-m", :type =>
+  :array, :default =>  [ "mustache/"], 
+  :desc => "Array of directory paths (ending with slash '/' char e.g. 'mustache/') of Gem names (optionally with version constraint e.g. 'aws-must-templates,~>0.0.1') holding mustache templates"
 
   # ------------------------------------------------------------------
   # action 'doc'
@@ -66,6 +68,7 @@ class App < Thor
   desc "doc", "Extract template documentation"
 
   shared_options :template_path
+  shared_options :mustaches
   shared_options :template_gem
 
   option :fold_on, :aliases => "-n", :type => :string, 
@@ -108,7 +111,7 @@ LONGDESC
 
   def doc( template_name=DEAFAULT_TEMPLATE )
 
-    @logger.info( "#{__FILE__}.#{__method__} starting" )
+    @logger.info( "#{__method__} starting" )
 
     
     my_options = option_adjust_common( options )
@@ -174,6 +177,7 @@ LONGDESC
   desc "gen <yaml_file> [<template_name>]", "Generate CloudFormation JSON template"
 
   shared_options :template_path
+  shared_options :mustaches
   shared_options :template_gem
 
   long_desc <<-LONGDESC
@@ -201,33 +205,90 @@ LONGDESC
 
     def option_adjust_common( options ) 
 
-      raise "No optios give" unless options
-      ret = options
-      ret = option_template_gem_to_template_path( options )  if options[:template_gem]
+      raise "No optios given" unless options
+
+      # options readonly - dup allows modification
+      options = options.dup
+
+      @logger.debug( "#{__method__}, options '#{options}'" )
+
+      options = option_template_gem_to_template_path( options )  if options[:template_gem]
+      options = option_template_path( options )  if options[:template_path]
+
+      warn <<-EOS   unless options[:mustaches] && options[:mustaches].any?
+
+      Deprecation!!
+
+           Use option -m to define location of templates
+
+           Options -g and -t deprecated
+      EOS
+
+      options = option_mustaches_to_template_paths( options )  if options[:mustaches]
       
-      return ret
+      return options
     end
 
+    # set `options[:template_path]` to an array of local directory
+    # paths parsed from list of options[:mustaches]
+    def option_mustaches_to_template_paths( options ) 
 
+      # init, NB: plural
+      options[:template_paths] = []
+
+      options[:mustaches].each do |mustache_def|
+        if mustache_def[-1] == '/' then
+          # directory with last char removed
+          options[:template_paths] << mustache_def[0..-1]
+        else
+          # gemname 
+          options[:template_paths] << gem_name_to_template_path( mustache_def )
+        end
+          
+      end
+      return options
+    end
+
+    # overrides -m option
+    def option_template_path( options ) 
+      # overrides -m option
+      options[:mustaches] = []
+      return options
+    end
+
+    # set 'options[:template_path]' to mustache directory withing
+    # `options[:template_gem]` Gem
     def option_template_gem_to_template_path( options ) 
+
+      # overrides -m option
+      options[:mustaches] = []
+
+      options[:template_path] = gem_name_to_template_path( options[:template_gem] )
+      
+      return options
+
+    end
+
+    # map 'gem_name_and_spec' to path of mustache templates
+    def gem_name_to_template_path( gem_name_and_spec ) 
 
       # The version requirements are optional.
       # You can also specify multiple version requirements, just append more at the end
-      gem_spec = options[:template_gem].split(',')
+      gem_spec = gem_name_and_spec.split(',')
       gem_name, *gem_ver_reqs = gem_spec[0], gem_spec[1]
+      @logger.debug( "#{__method__}, gem_name=#{gem_name}, *gem_ver_reqs=#{gem_ver_reqs}" )
       gdep = Gem::Dependency.new(gem_name, *gem_ver_reqs)
       # find latest that satisifies
       found_gspec = gdep.matching_specs.sort_by(&:version).last
+      @logger.debug( "#{__method__}, found_gspec=#{found_gspec}" )
       # instead of using Gem::Dependency, you can also do:
       # Gem::Specification.find_all_by_name(gem_name, *gem_ver_reqs)
 
       if found_gspec
-        # unfreeze options
-        @logger.debug( "#{__FILE__}.#{__method__}, Requirement '#{gdep}' already satisfied by #{found_gspec.name}-#{found_gspec.version}" )
-        @logger.debug( "#{__FILE__}.#{__method__}, old-options=#{options}" )
-        options = options.dup
-        options[:template_path] = "#{found_gspec.gem_dir}/mustache"
-        @logger.debug( "#{__FILE__}.#{__method__}, new-options=#{options}" )
+        @logger.debug( "#{__method__}, Requirement '#{gdep}' already satisfied by #{found_gspec.name}-#{found_gspec.version}" )
+        @logger.debug( "#{__method__}, old-options=#{options}" )
+        template_path = "#{found_gspec.gem_dir}/mustache"
+        @logger.debug( "#{__method__}, template_path=#{template_path}" )
       else
         #puts "Requirement '#{gdep}' not satisfied; installing..."
         # ver_args = gdep.requirements_list.map{|s| ['-v', s] }.flatten
@@ -236,7 +297,7 @@ LONGDESC
         raise "Could not find gem '#{gdep}' - try 'gem install #{gdep}'"
       end
       
-      return options
+      return template_path
 
     end