+++ lib/aws-must/template.rb
@@ -12,10 +12,6 @@ module AwsMust
     # ------------------------------------------------------------------
     # Attributes
 
-    # static
-    @@template_path = nil            # directory where templates stored
-    @@template_extension = "mustache"# type part in filename
-
     # instance 
     attr_writer :partials            # f: partial-name --> template string
     attr_writer :templates           # f: template-name --> template string
@@ -25,10 +21,18 @@ module AwsMust
     
     def initialize( options={} )
       @logger = getLogger( PROGNAME, options )
-      @logger.info( "#{__FILE__}.#{__method__} created" )
-      @logger.debug( "#{__FILE__}.#{__method__}, options='#{options}" )
+      @logger.info( "#{__method__} created" )
+      @logger.debug( "#{__method__}, options='#{options}" )
+
+      @template_extension = "mustache"# type part in filename
+
       # for mustache templates
-      @@template_path = options[:template_path] if options[:template_path]
+      if options[:template_paths] && options[:template_paths].any? then
+        @template_paths = options[:template_paths]
+      else
+        @template_paths = ( options[:template_path].respond_to?(:each) ? options[:template_path] : [ options[:template_path]  ] )
+      end
+
     end
 
     # ------------------------------------------------------------------
@@ -51,55 +55,81 @@ module AwsMust
 
     # method used by mustache framework - delegate to 'get_partial'
     def partial(name)
-      @logger.debug( "#{__FILE__}.#{__method__} name=#{name}" )
-      # File.read("#{template_path}/#{name}.#{template_extension}")
+      @logger.debug( "#{__method__} name=#{name}" )
       get_partial( name )
     end
 
     # cache @partials - for easier extension
     def get_partial( name ) 
-      @logger.debug( "#{__FILE__}.#{__method__} name=#{name}" )
+      @logger.debug( "#{__method__} name=#{name}" )
       return @partials[name] if @partials && @partials[name]
 
-      partial_file = "#{@@template_path}/#{name}.#{template_extension}"
-      @logger.info( "#{__FILE__}.#{__method__} read partial_file=#{partial_file}" )
+      partial_file = get_template_filepath( name )
+      @logger.info( "#{__method__} read partial_file=#{partial_file}" )
       File.read( partial_file )
     end
 
     # hide @templates - for easier extension
     def get_template( name ) 
 
-      @logger.debug( "#{__FILE__}.#{__method__} name=#{name}" )
-      return @templates[name] if @templates && @templates[name]
+      template_file = get_template_filepath( name )
+      @logger.info( "#{__method__} read template_file=#{template_file}" )
+      File.read( template_file )
 
-      if !  File.exists?( @@template_path ) then
-        raise <<-eos
+    end # def get_template( name ) 
 
 
-          No such directory '#{@@template_path}'.
+    # return path to an existing template file name 
+    private
+
+    def get_template_filepath( name ) 
+
+      @template_paths.each do |directory| 
+
+        template_path = get_template_filepath_in_directory( directory, name ) 
+        
+        return template_path if File.exists?( template_path ) 
+
+      end # each 
+
+
+      # could not find 
+      
+      raise <<-eos
+
+          No such template '#{name}' found in directories #{@template_paths.join(", ")}
  
-          Use opition -t to point to an existing directory.
+          Use opition -g list directories or Gems, where file '#{name}.#{@template_extension}' can be located.
         
         eos
-      end
-
-      template_path = "#{@@template_path}/#{name}.#{@@template_extension}"
-      @logger.info( "#{__FILE__}.#{__method__} read template_path=#{template_path}" )
-
-      if !  File.exists?( template_path ) then
-        raise <<-eos
-
-          No such file  '#{template_path}'.
- 
-          Use opition -t to point to a directory, which contains file '#{name}.#{@@template_extension}'
-
-        
-        eos
-      end
 
 
-      File.read( template_path )
     end
+
+    # return path to 'template_file' in an existing 'directory'
+    def get_template_filepath_in_directory( directory, template_file ) 
+      @logger.debug( "#{__method__} directory=#{directory}, template_file=#{template_file}" )
+
+      if ! File.exists?( directory ) then
+        raise <<-eos
+
+          No such directory '#{directory}'.
+ 
+          Option -g should list
+          - existing paths OR
+          - Gems which includes 'mustache' directory
+        
+        eos
+
+      end
+
+      template_path = "#{directory}/#{template_file}.#{@template_extension}"
+      @logger.info( "#{__method__} read template_path=#{template_path}" )
+      
+      return template_path
+
+    end
+
 
   end # class